정말 머저리같은 malloc 호출을 수정함 내가 봐도 소름 돋는 ㅄ짓이었음

+++ console_project/structs.h
@@ -83,7 +83,7 @@ struct IntVec2
 	}
 	operator COORD() const
 	{
-		COORD coord = { x, y };
+		COORD coord = { static_cast<short>(x), static_cast<short>(y) };
 		return coord;
 	}
 	static int Dot(IntVec2 left, IntVec2 right)
